+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* EXPERIMENTAL PERMISSION CACHING IN REFERENCE DATA LOOKUP
+
 * EXPERIMENTAL REFERENCE DATA LOOKUP WITHOUT PERMISSION CHECKING.
 
 ## [v5.3.0-beta.2] - 2018-04-30